feat(desktop): unify the sidebar into one chats-first rail
